migrate ProxyLocation to skyvern.schemas.runs
