+++ skyvern/agent/client.py
@@ -6,8 +6,8 @@ import httpx
 from skyvern.config import settings
 from skyvern.exceptions import SkyvernClientException
 from skyvern.forge.sdk.schemas.task_runs import TaskRunResponse
-from skyvern.forge.sdk.schemas.tasks import ProxyLocation
 from skyvern.forge.sdk.workflow.models.workflow import RunWorkflowResponse, WorkflowRunResponse
+from skyvern.schemas.runs import ProxyLocation
 
 
 class RunEngine(StrEnum):
