+++ skyvern/forge/sdk/workflow/models/workflow.py
@@ -7,10 +7,10 @@ from pydantic import BaseModel, field_validator
 from skyvern.forge.sdk.core.validators import validate_url
 from skyvern.forge.sdk.schemas.files import FileInfo
 from skyvern.forge.sdk.schemas.task_v2 import TaskV2
-from skyvern.forge.sdk.schemas.tasks import ProxyLocation
 from skyvern.forge.sdk.workflow.exceptions import WorkflowDefinitionHasDuplicateBlockLabels
 from skyvern.forge.sdk.workflow.models.block import BlockTypeVar
 from skyvern.forge.sdk.workflow.models.parameter import PARAMETER_TYPE
+from skyvern.schemas.runs import ProxyLocation
 
 
 class WorkflowRequestBody(BaseModel):
